# api/openapi.yaml
openapi: 3.1.0
info:
  title: MCPDoll Control Plane API
  version: 0.1.0
  summary: The single contract behind MCPDoll's API, CLI, and console.
  description: |
    This document is the source of truth for every MCPDoll operation. The Go
    server interfaces, the Go client the CLI uses, the TypeScript client the
    console uses, and `docs/api/` are all generated from it.

    Two rules make that more than a slogan:

    - **The CLI is an API client and never touches the database.** Anything the
      CLI can do, an external integrator can do, because the CLI has no
      privileged path.
    - **Every operation declares its surfaces** in `x-mcpdoll-surfaces`, and
      `make parity` fails the build if an operation lacks a CLI command or a UI
      route, or if a CLI or UI binding names an operation that does not exist.

    See `docs/adr/0004-api-first-tri-surface.md`.

    ## What this version covers

    The registry is a reviewable YAML document rather than database rows, and
    the control plane's durable side is not built (see `docs/deferred.md`). So
    the operations here are the ones that are genuinely implemented: reading the
    registry, inspecting and verifying snapshots, and observing a live data
    plane. Registry *mutation* through the API arrives with the database.

  license:
    name: MIT
    # SPDX, so the declaration is machine-readable and matches ./LICENSE rather
    # than only agreeing with it by eye.
    identifier: MIT
servers:
...
paths:
...
  /api/v1/schedules/{jobType}:
    get:
      operationId: getSchedule
      summary: Read one piece of timed work
      tags: [schedules]
      x-mcpdoll-surfaces:
        cli: mcpdoll schedules show
        ui: /schedules/:jobType
      parameters:
        - { name: jobType, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: The schedule
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Schedule" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateSchedule
      summary: Retune or switch off one piece of timed work
      description: |
        Changing a cadence takes effect without a restart, and re-arms the next
        run from now — otherwise lengthening an interval would appear not to
        have taken until the old one elapsed.
      tags: [schedules]
      x-mcpdoll-surfaces:
        cli: mcpdoll schedules set
        ui: /schedules/:jobType/edit
      parameters:
        - { name: jobType, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateScheduleRequest" }
      responses:
        "200":
          description: The updated schedule
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Schedule" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/schedules/{jobType}:run:
    post:
      operationId: runScheduleNow
      summary: Bring a schedule forward so it runs on the next tick
      description: |
        Does not run the job inline. It moves the next run to now and the
        scheduler picks it up, through the same claim and the same outcome
        recording as any other run — a second execution path would be a second
        way for the job to happen, and the two would drift.
      tags: [schedules]
      x-mcpdoll-surfaces:
        cli: mcpdoll schedules run
        ui: /schedules/:jobType/run
      parameters:
        - { name: jobType, in: path, required: true, schema: { type: string } }
      responses:
        "202":
          description: The schedule, now due
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Schedule" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
...
  responses:
    NotFound:
      description: The resource does not exist
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            Stable machine-readable token. Clients switch on this; the message
            is for humans and may be reworded.
          enum:
            - invalid_request
            - not_found
            - forbidden
            - validation_failed
            - upstream_unavailable
            - internal_error
        message:
          type: string
          description: One sentence, written for whoever has to fix the problem.
        problems:
          type: array
          items: { type: string }
          description: |
            Every distinct failure. Validation returns all of them at once: a
            document with six errors should take one round trip to fix.
...
    Schedule:
      type: object
      required: [job_type, name, kind, spec, enabled, system]
      properties:
        job_type:
          type: string
          description: Dispatch key and stable identity; a retuned cadence hangs off it.
        name: { type: string }
        kind:
          type: string
          enum: [interval]
          description: |
            Only `interval` today. Cron's finest granularity is a minute and the
            revocation heartbeat runs every thirty seconds, so the cadence that
            matters most here is one cron cannot express.
        spec:
          type: string
          description: A Go duration for an interval schedule — "30s", "1m".
        enabled: { type: boolean }
        system:
          type: boolean
          description: |
            Cannot be deleted, because nothing would recreate it. Still
            retunable and still possible to switch off.
        next_run_at: { type: string, format: date-time }
        last_run_at: { type: string, format: date-time }
        last_error:
          type: string
          description: Why the last run failed, absent when it succeeded.
        last_duration_ms: { type: integer }
...
    UpdateScheduleRequest:
      type: object
      properties:
        spec:
          type: string
          description: New cadence as a Go duration. Omitted leaves it alone.
        enabled:
          type: boolean
          description: Omitted leaves it alone, so either can change without the other.
...