# internal/controlplane/store/dbgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package dbgen

import (
	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type ApiKey struct {
	ID         uuid.UUID
	UserID     uuid.UUID
	Name       string
	Prefix     string
	Hash       string
	CreatedAt  pgtype.Timestamptz
	LastUsedAt pgtype.Timestamptz
	ExpiresAt  pgtype.Timestamptz
	RevokedAt  pgtype.Timestamptz
	TenantID   uuid.UUID
}

type ApiKeyGrant struct {
	ID       uuid.UUID
	ApiKeyID uuid.UUID
	Role     string
	Scope    string
}

type AuthSetting struct {
	ID          bool
	SignupMode  string
	DefaultRole *string
	UpdatedAt   pgtype.Timestamptz
}

type Grant struct {
	ID        uuid.UUID
	UserID    uuid.UUID
	Role      string
	Scope     string
	GrantedBy *uuid.UUID
	CreatedAt pgtype.Timestamptz
}

type IdentityProvider struct {
	ID          uuid.UUID
	Slug        string
	Kind        string
	DisplayName string
	Enabled     bool
	TenantID    *uuid.UUID
	Config      []byte
	CreatedAt   pgtype.Timestamptz
	UpdatedAt   pgtype.Timestamptz
}

type Revocation struct {
	ID           int64
	PrincipalID  uuid.UUID
	Kind         string
	UserID       *uuid.UUID
	Reason       *string
	RevokedAt    pgtype.Timestamptz
	SupersededBy *int64
}

type RevocationState struct {
	ID               bool
	Version          int64
	PrunedThrough    int64
	UpdatedAt        pgtype.Timestamptz
	PrincipalVersion int64
}

type RolePermission struct {
	Role       string
	Permission string
}

type Schedule struct {
	ID             uuid.UUID
	JobType        string
	Name           string
	Kind           string
	Spec           string
	Enabled        bool
	System         bool
	NextRunAt      pgtype.Timestamptz
	LastRunAt      pgtype.Timestamptz
	LastError      *string
	LastDurationMs *int32
	CreatedAt      pgtype.Timestamptz
	UpdatedAt      pgtype.Timestamptz
}

type Session struct {
	ID         uuid.UUID
	UserID     uuid.UUID
	Prefix     string
	Hash       string
	UserAgent  *string
	Ip         *string
	CreatedAt  pgtype.Timestamptz
	LastSeenAt pgtype.Timestamptz
	ExpiresAt  pgtype.Timestamptz
	RevokedAt  pgtype.Timestamptz
}

type Tenant struct {
	ID        uuid.UUID
	Slug      string
	Name      string
	Status    string
	Metadata  []byte
	CreatedAt pgtype.Timestamptz
	UpdatedAt pgtype.Timestamptz
}

type User struct {
	ID           uuid.UUID
	Email        string
	DisplayName  *string
	PasswordHash *string
	Status       string
	CreatedAt    pgtype.Timestamptz
	UpdatedAt    pgtype.Timestamptz
}

type UserIdentity struct {
	ID        uuid.UUID
	UserID    uuid.UUID
	Provider  string
	Subject   string
	Email     *string
	CreatedAt pgtype.Timestamptz
}